move src.utils.statistics to src.document.statistics (the measures) and to src.utils_statistic_utils (the helper functions)
