+++ configs/dataset/_collect_statistics.yaml
@@ -8,45 +8,45 @@ collect_statistics:
   _enabled_: true
 
 count_documents:
-  _processor_: src.utils.statistics.generic_collect_statistics
+  _processor_: src.utils.collect_statistics
   _enabled_: ${dataset.collect_statistics._enabled_}
   format: markdown
   title: number of documents
   aggregate_functions:
     - sum
   measure:
-    _target_: src.utils.statistics.DummyCounter
+    _target_: src.document.statistics.DummyCounter
 
 collect_label_counts:
-  _processor_: src.utils.statistics.generic_collect_statistics
+  _processor_: src.utils.collect_statistics
   _enabled_: ${dataset.collect_statistics._enabled_}
   format: plot
   title: number of ${dataset.collect_statistics.labeled_field} labels
   aggregate_functions:
     - mean
   measure:
-    _target_: src.utils.statistics.LabelCounter
+    _target_: src.document.statistics.LabelCounter
     field: ${dataset.collect_statistics.labeled_field}
 
 collect_label_counts_hist:
-  _processor_: src.utils.statistics.generic_collect_statistics
+  _processor_: src.utils.collect_statistics
   _enabled_: ${dataset.collect_statistics._enabled_}
   format: plot
   title: ${dataset.collect_statistics.labeled_field} labels distribution
   measure:
-    _target_: src.utils.statistics.LabelCounter
+    _target_: src.document.statistics.LabelCounter
     field: ${dataset.collect_statistics.labeled_field}
 
 collect_field_lengths:
-  _processor_: src.utils.statistics.generic_collect_statistics
+  _processor_: src.utils.collect_statistics
   _enabled_: ${dataset.collect_statistics._enabled_}
   format: markdown
   title: number of characters
   measure:
-    _target_: src.utils.statistics.DocumentFieldLengthCounter
+    _target_: src.document.statistics.DocumentFieldLengthCounter
     field: ${dataset.collect_statistics.text_field}
 #collect_span_lengths:
-#  _processor_: src.utils.statistics.generic_collect_statistics
+#  _processor_: src.utils.collect_statistics
 #  _enabled_: ${dataset.collect_statistics._enabled_}
 #  format: plot
 #  title: ${dataset.collect_statistics.labeled_span_field} lengths distribution
