+++ src/utils/__init__.py
@@ -1,4 +1,5 @@
 from .config_utils import execute_pipeline, instantiate_dict_entries, prepare_omegaconf
 from .logging_utils import close_loggers, get_pylogger, log_hyperparameters
 from .rich_utils import enforce_tags, print_config_tree
+from .statistic_utils import collect_statistics
 from .task_utils import extras, save_file, task_wrapper
